% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouncin.R
\name{bouncin}
\alias{bouncin}
\title{bouncin}
\arguments{
\item{n_point}{number of desired points for coordinates}

\item{w}{width of canvas}

\item{h}{height of canvas}
}
\value{
data frame of coordinates
}
\description{
determine coordinate path of object for a parabolic bounce
}
\examples{
bouncin(17,5000,3000)
bouncin(30,720,720)
}
