% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncbpm.R
\name{syncbpm}
\alias{syncbpm}
\title{syncbpm}
\usage{
syncbpm(x, y, z, n)
}
\arguments{
\item{x}{duration of animation loop}

\item{y}{BPM of song}

\item{z}{beats per animation loop}

\item{n}{FPS of animation}
}
\value{
prompt containing percentage
}
\description{
determine bpm of song to match fps of frame by frame animation for rhythmic visual
}
\examples{
syncbpm(15,177,2,24)
}
