% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncfps.R
\name{syncfps}
\alias{syncfps}
\title{syncfps}
\usage{
syncfps(x, y, z, n)
}
\arguments{
\item{x}{duration of animation loop}

\item{y}{BPM of song}

\item{z}{beats per animation loop}

\item{n}{FPS of animation}
}
\value{
prompt containing percentage
}
\description{
determine speed of frame by frame animation to match bpm of song for a rhythmic visual
}
\examples{
syncfps(15,177,2,24)
}
